+++ Assets/Scripts/PlayerStats.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 public class PlayerStats
 {
     // Add more levelcaps and actual logic for it
-    static int[] xpCaps = { 0, 200, 300, 400, 500, 600 };
+    static float[] xpCaps = { 0, 200, 300, 400, 500, 600 };
     static float[] baseDamagePerLevel = { 0, 6f, 7f, 8f, 9f, 10f };
     static float[] baseMaxHpPerLevel = { 0, 50f, 60f, 65f, 75f, 80f };
     //Level specific stuff ends
@@ -29,9 +29,9 @@ public class PlayerStats
     }
     public static float hp = maxHp;
     public static float damageReductionMultiplier = 1f;
-	public static int xp = 0;
+	public static float xp = 0;
     
-	public static int maxLevel{
+	public static float maxLevel{
         get{
             return xpCaps.Length;
         }
@@ -39,7 +39,7 @@ public class PlayerStats
 	public static float attackSpeed = 1.1f;
     public static string playerName = "Hero";
     public static Entity target;
-    public static int xpToNextLvl{
+    public static float xpToNextLvl{
         get{
             return xpCaps[level];
         }
@@ -52,7 +52,7 @@ public class PlayerStats
         hp = maxHp;
     }
     
-    public static void GainXp(int amount){
+    public static void GainXp(float amount){
         if (level == maxLevel){
             xp = 0;
             return;
